Refactor joint application review notifications
