+++ src/main/java/uk/gov/hmcts/divorce/citizen/notification/JointApplicationApprovedReminder.java
@@ -1,0 +1,45 @@
+package uk.gov.hmcts.divorce.citizen.notification;
+
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Component;
+import uk.gov.hmcts.divorce.divorcecase.model.CaseData;
+import uk.gov.hmcts.divorce.notification.ApplicantNotification;
+import uk.gov.hmcts.divorce.notification.CommonContent;
+import uk.gov.hmcts.divorce.notification.NotificationService;
+
+import java.util.Map;
+
+import static uk.gov.hmcts.divorce.citizen.notification.Applicant2ApprovedNotification.PAYS_FEES;
+import static uk.gov.hmcts.divorce.notification.CommonContent.IS_REMINDER;
+import static uk.gov.hmcts.divorce.notification.CommonContent.NO;
+import static uk.gov.hmcts.divorce.notification.CommonContent.YES;
+import static uk.gov.hmcts.divorce.notification.EmailTemplateName.JOINT_APPLICATION_APPROVED_APPLICANT1_REMINDER;
+
+@Component
+@Slf4j
+public class JointApplicationApprovedReminder implements ApplicantNotification {
+
+    @Autowired
+    private NotificationService notificationService;
+
+    @Autowired
+    private CommonContent commonContent;
+
+    @Override
+    public void sendToApplicant1(final CaseData caseData, final Long id) {
+        log.info("Sending notification to applicant 1 to notify them that applicant 2 has reviewed the application: {}", id);
+
+        Map<String, String> templateVars =
+            commonContent.mainTemplateVars(caseData, id, caseData.getApplicant1(), caseData.getApplicant2());
+        templateVars.put(PAYS_FEES, caseData.getApplication().isHelpWithFeesApplication() ? NO : YES);
+        templateVars.put(IS_REMINDER, YES);
+
+        notificationService.sendEmail(
+            caseData.getApplicant1().getEmail(),
+            JOINT_APPLICATION_APPROVED_APPLICANT1_REMINDER,
+            templateVars,
+            caseData.getApplicant1().getLanguagePreference()
+        );
+    }
+}
