+++ src/main/java/uk/gov/hmcts/divorce/citizen/notification/JointApplicationNotReviewedNotification.java
@@ -4,23 +4,19 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import uk.gov.hmcts.divorce.divorcecase.model.CaseData;
+import uk.gov.hmcts.divorce.notification.ApplicantNotification;
 import uk.gov.hmcts.divorce.notification.CommonContent;
 import uk.gov.hmcts.divorce.notification.NotificationService;
 
 import java.util.Map;
 
-import static uk.gov.hmcts.divorce.citizen.notification.Applicant2ApprovedNotification.PAYS_FEES;
-import static uk.gov.hmcts.divorce.notification.CommonContent.IS_REMINDER;
-import static uk.gov.hmcts.divorce.notification.CommonContent.NO;
 import static uk.gov.hmcts.divorce.notification.CommonContent.REVIEW_DEADLINE_DATE;
-import static uk.gov.hmcts.divorce.notification.CommonContent.YES;
-import static uk.gov.hmcts.divorce.notification.EmailTemplateName.JOINT_APPLICATION_APPROVED_APPLICANT1_REMINDER;
 import static uk.gov.hmcts.divorce.notification.EmailTemplateName.JOINT_APPLICATION_OVERDUE;
 import static uk.gov.hmcts.divorce.notification.FormatUtil.DATE_TIME_FORMATTER;
 
 @Component
 @Slf4j
-public class JointApplicationOverdueNotification {
+public class JointApplicationNotReviewedNotification implements ApplicantNotification {
 
     @Autowired
     private NotificationService notificationService;
@@ -28,7 +24,8 @@ public class JointApplicationOverdueNotification {
     @Autowired
     private CommonContent commonContent;
 
-    public void sendApplicationNotReviewedEmail(CaseData caseData, Long id) {
+    @Override
+    public void sendToApplicant1(final CaseData caseData, final Long id) {
         log.info("Sending notification to applicant 1 to notify them of overdue joint application: {}", id);
 
         Map<String, String> templateVars =
@@ -38,22 +35,6 @@ public class JointApplicationOverdueNotification {
         notificationService.sendEmail(
             caseData.getApplicant1().getEmail(),
             JOINT_APPLICATION_OVERDUE,
-            templateVars,
-            caseData.getApplicant1().getLanguagePreference()
-        );
-    }
-
-    public void sendApplicationApprovedReminderToApplicant1(CaseData caseData, Long id) {
-        log.info("Sending notification to applicant 1 to notify them that applicant 2 has reviewed the application: {}", id);
-
-        Map<String, String> templateVars =
-            commonContent.mainTemplateVars(caseData, id, caseData.getApplicant1(), caseData.getApplicant2());
-        templateVars.put(PAYS_FEES, caseData.getApplication().isHelpWithFeesApplication() ? NO : YES);
-        templateVars.put(IS_REMINDER, YES);
-
-        notificationService.sendEmail(
-            caseData.getApplicant1().getEmail(),
-            JOINT_APPLICATION_APPROVED_APPLICANT1_REMINDER,
             templateVars,
             caseData.getApplicant1().getLanguagePreference()
         );
