+++ src/test/java/uk/gov/hmcts/divorce/citizen/notification/JointApplicationApprovedReminderTest.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
-import org.springframework.test.context.junit.jupiter.SpringExtension;
+import org.mockito.junit.jupiter.MockitoExtension;
 import uk.gov.hmcts.ccd.sdk.type.YesOrNo;
 import uk.gov.hmcts.divorce.divorcecase.model.CaseData;
 import uk.gov.hmcts.divorce.divorcecase.model.DivorceOrDissolution;
@@ -29,18 +29,15 @@ import static uk.gov.hmcts.divorce.notification.CommonContent.IS_DISSOLUTION;
 import static uk.gov.hmcts.divorce.notification.CommonContent.IS_DIVORCE;
 import static uk.gov.hmcts.divorce.notification.CommonContent.IS_REMINDER;
 import static uk.gov.hmcts.divorce.notification.CommonContent.NO;
-import static uk.gov.hmcts.divorce.notification.CommonContent.REVIEW_DEADLINE_DATE;
 import static uk.gov.hmcts.divorce.notification.CommonContent.YES;
 import static uk.gov.hmcts.divorce.notification.EmailTemplateName.JOINT_APPLICATION_APPROVED_APPLICANT1_REMINDER;
-import static uk.gov.hmcts.divorce.notification.EmailTemplateName.JOINT_APPLICATION_OVERDUE;
-import static uk.gov.hmcts.divorce.notification.FormatUtil.DATE_TIME_FORMATTER;
 import static uk.gov.hmcts.divorce.testutil.TestConstants.TEST_USER_EMAIL;
 import static uk.gov.hmcts.divorce.testutil.TestDataHelper.caseData;
 import static uk.gov.hmcts.divorce.testutil.TestDataHelper.getApplicant;
 import static uk.gov.hmcts.divorce.testutil.TestDataHelper.getMainTemplateVars;
 
-@ExtendWith(SpringExtension.class)
-public class JointApplicationOverdueNotificationTest {
+@ExtendWith(MockitoExtension.class)
+class JointApplicationApprovedReminderTest {
 
     @Mock
     private NotificationService notificationService;
@@ -49,53 +46,7 @@ public class JointApplicationOverdueNotificationTest {
     private CommonContent commonContent;
 
     @InjectMocks
-    private JointApplicationOverdueNotification notification;
-
-    @Test
-    void shouldSendEmailToApplicant1WithDivorceContent() {
-        CaseData data = caseData();
-        data.setDueDate(LocalDate.now());
-        data.setApplicant2(getApplicant(Gender.FEMALE));
-        when(commonContent.mainTemplateVars(data, 1234567890123456L, data.getApplicant1(), data.getApplicant2()))
-            .thenReturn(getMainTemplateVars());
-
-        notification.sendApplicationNotReviewedEmail(data, 1234567890123456L);
-
-        verify(notificationService).sendEmail(
-            eq(TEST_USER_EMAIL),
-            eq(JOINT_APPLICATION_OVERDUE),
-            argThat(allOf(
-                hasEntry(REVIEW_DEADLINE_DATE, LocalDate.now().format(DATE_TIME_FORMATTER))
-            )),
-            eq(ENGLISH)
-        );
-        verify(commonContent).mainTemplateVars(data, 1234567890123456L, data.getApplicant1(), data.getApplicant2());
-    }
-
-    @Test
-    void shouldSendEmailToApplicant1WithDissolutionContent() {
-        CaseData data = caseData();
-        data.setDueDate(LocalDate.now());
-        data.setDivorceOrDissolution(DivorceOrDissolution.DISSOLUTION);
-        data.setApplicant2(getApplicant(Gender.MALE));
-        final Map<String, String> templateVars = getMainTemplateVars();
-        templateVars.putAll(Map.of(IS_DISSOLUTION, YES, IS_DIVORCE, NO));
-        when(commonContent.mainTemplateVars(data, 1234567890123456L, data.getApplicant1(), data.getApplicant2())).thenReturn(templateVars);
-
-        notification.sendApplicationNotReviewedEmail(data, 1234567890123456L);
-
-        verify(notificationService).sendEmail(
-            eq(TEST_USER_EMAIL),
-            eq(JOINT_APPLICATION_OVERDUE),
-            argThat(allOf(
-                hasEntry(IS_DISSOLUTION, YES),
-                hasEntry(IS_DIVORCE, NO),
-                hasEntry(REVIEW_DEADLINE_DATE, LocalDate.now().format(DATE_TIME_FORMATTER))
-            )),
-            eq(ENGLISH)
-        );
-        verify(commonContent).mainTemplateVars(data, 1234567890123456L, data.getApplicant1(), data.getApplicant2());
-    }
+    private JointApplicationApprovedReminder jointApplicationApprovedReminder;
 
     @Test
     void shouldSendReminderEmailToApplicant1WithDivorceAndPaymentContent() {
@@ -105,7 +56,7 @@ public class JointApplicationOverdueNotificationTest {
         when(commonContent.mainTemplateVars(data, 1234567890123456L, data.getApplicant1(), data.getApplicant2()))
             .thenReturn(getMainTemplateVars());
 
-        notification.sendApplicationApprovedReminderToApplicant1(data, 1234567890123456L);
+        jointApplicationApprovedReminder.sendToApplicant1(data, 1234567890123456L);
 
         verify(notificationService).sendEmail(
             eq(TEST_USER_EMAIL),
@@ -129,7 +80,7 @@ public class JointApplicationOverdueNotificationTest {
         when(commonContent.mainTemplateVars(data, 1234567890123456L, data.getApplicant1(), data.getApplicant2()))
             .thenReturn(getMainTemplateVars());
 
-        notification.sendApplicationApprovedReminderToApplicant1(data, 1234567890123456L);
+        jointApplicationApprovedReminder.sendToApplicant1(data, 1234567890123456L);
 
         verify(notificationService).sendEmail(
             eq(TEST_USER_EMAIL),
@@ -152,7 +103,7 @@ public class JointApplicationOverdueNotificationTest {
         final Map<String, String> templateVars = getMainTemplateVars();
         when(commonContent.mainTemplateVars(data, 1234567890123456L, data.getApplicant1(), data.getApplicant2())).thenReturn(templateVars);
 
-        notification.sendApplicationApprovedReminderToApplicant1(data, 1234567890123456L);
+        jointApplicationApprovedReminder.sendToApplicant1(data, 1234567890123456L);
 
         verify(notificationService).sendEmail(
             eq(TEST_USER_EMAIL),
@@ -178,7 +129,7 @@ public class JointApplicationOverdueNotificationTest {
         templateVars.putAll(Map.of(IS_DISSOLUTION, YES, IS_DIVORCE, NO));
         when(commonContent.mainTemplateVars(data, 1234567890123456L, data.getApplicant1(), data.getApplicant2())).thenReturn(templateVars);
 
-        notification.sendApplicationApprovedReminderToApplicant1(data, 1234567890123456L);
+        jointApplicationApprovedReminder.sendToApplicant1(data, 1234567890123456L);
 
         verify(notificationService).sendEmail(
             eq(TEST_USER_EMAIL),
